Fix for bug #3132.
Renamed the executable <bindir>/snarf to <bindir>/gorge, so it does
not conflict with the executable shipped with the snarf project (see
http://www.xach.com/snarf/).

Added support for a system-wide rc/configuration file.  Ship a default file
that gets installed at <sysconfdir>/spew.conf.  

Added the --no-rcfiles option which is used to ignore the standard rc files.